MovieVerse-AI/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
pandas==2.2.2
numpy==1.26.4
scipy==1.12.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pymysql==1.1.0
pymongo==4.7.2
kafka-python==2.0.2
lightfm==1.17
scikit-learn==1.4.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
mlflow==2.12.1
feast==0.38.1
redis==5.0.4
prometheus-client==0.20.0
joblib==1.4.2
boto3==1.34.106
transformers==4.41.2
torch==2.3.1
torchvision==0.18.1
pillow==10.3.0
sentencepiece==0.2.0
httpx==0.27.0
